pyproject.toml
[tool.poetry]
name = "wappu_spiriter"
version = "0.1.0"
description = ""
authors = ["Mikael Siidorow <mikael.siidorow@aalto.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
python-telegram-bot = "^21.0.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
openai = "^1.16.2"
pillow = "^10.3.0"
matplotlib = "^3.8.4"
more-itertools = "^10.2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
poethepoet = "^0.25.0"
ruff = "^0.3.5"
watchfiles = "^0.21.0"

[tool.poe.tasks]
format = [{ cmd = "ruff format ." }, { cmd = "ruff check . --fix" }]
check = "mypy ."
lint = "ruff check ."
start = "python -m wappu_spiriter"
dev = "watchfiles \"poe start\" wappu_spiriter"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
